tools/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tdl-validator"
version = "1.0.0"
description = "Validation tools for TDL (Tutor Description Language) files"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Pedro A. Pernías Peco", email = "p.pernias@gmail.com"}
]
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.17",
]

[project.urls]
Homepage = "https://github.com/ppernias/tdl-spec"
Documentation = "https://github.com/ppernias/tdl-spec#readme"
Repository = "https://github.com/ppernias/tdl-spec"
Issues = "https://github.com/ppernias/tdl-spec/issues"

[project.scripts]
tdl-validate = "tdl_validator.cli:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
"*" = ["*.json", "*.yaml"]
